Add TODOs for Cleanup and Implementations in AbortController
- Added TODO comments in `AbortController.h` for future cleanup and implementation of the `AbortSignal` struct.
- Deleted the `Protocol__DT.hpp` file to streamline the codebase and remove unused components, aligning with ongoing efforts to modernize the protocol and improve maintainability.
